requirements.txt
streamlit==1.31.1
pandas==2.2.0
openai==1.12.0
python-dotenv==1.0.1
httpx>=0.24.1
langchain>=0.1.0
langchain-core>=0.1.22
langchain-openai>=0.0.6
langchain-community>=0.0.24
faiss-cpu>=1.7.4
